refactor: Move function and useContext for state
