+++ src/Components/Buttons/ButtonNetwork.jsx
@@ -3,20 +3,14 @@ import { ErrorBoundary } from "react-error-boundary";
 import PropTypes from "prop-types";
 import { ErrorFallback } from "@Errors";
 
-const ButtonNetwork = ({ content, size, variant, classGeneral, event, params }) => {
+const ButtonNetwork = ({ content, event, params, ...args }) => {
   const [isShow, cardRef, network] = params;
   const paramsFunction = isShow ? { ref: cardRef, type: network } : null;
 
   return (
     <>
       <ErrorBoundary FallbackComponent={ErrorFallback}>
-        <Button
-          variant={variant}
-          size={size}
-          className={classGeneral}
-          onClick={() => event(paramsFunction)}
-          active
-        >
+        <Button {...args} onClick={() => event(paramsFunction)} active>
           {content}
         </Button>
       </ErrorBoundary>
@@ -26,9 +20,6 @@ const ButtonNetwork = ({ content, size, variant, classGeneral, event, params }) 
 
 ButtonNetwork.propTypes = {
   content: PropTypes.object.isRequired,
-  size: PropTypes.string.isRequired,
-  variant: PropTypes.string.isRequired,
-  classGeneral: PropTypes.string,
   event: PropTypes.func,
   params: PropTypes.arrayOf(PropTypes.any),
 };
