+++ src/Components/Cards/CardAuthor.jsx
@@ -1,12 +1,44 @@
+import { useContext, useCallback } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 import PropTypes from "prop-types";
 import { Card, Row, Col } from "react-bootstrap";
 import { ErrorFallback } from "@Errors";
 import { ButtonNetwork } from "@Components";
+import { ContextQuoteFetch, ContextQuote } from "@context";
+import { getRandomColor } from "@Services";
 
-const CardAuthor = ({ contentBtn, stylesBtn, contentCard, event, eventCard, ...props }) => {
+const CardAuthor = ({ contentBtn, stylesBtn, contentCard, ...props }) => {
+  const { fetchQuote, fetchQuoteAuthor } = useContext(ContextQuoteFetch);
+  const { updatePainting, updateColorRandom, updateColorRandomSpinner } = useContext(ContextQuote);
+
   const [classBtn, size, variant] = stylesBtn;
   const [title, subtitle] = contentCard;
+
+  const searchQuoteAboutAuthor = useCallback(
+    async (author) => {
+      // Get more quotes the same author
+      await fetchQuoteAuthor({
+        url: `https://api.quotable.io/quotes?author=${author}`,
+        method: "GET",
+      });
+
+      updatePainting(true);
+    },
+    [fetchQuoteAuthor, updatePainting],
+  );
+
+  const nextQuoteRandom = useCallback(() => {
+    // Get other quote random
+    fetchQuote({
+      url: "https://api.quotable.io/random",
+      method: "GET",
+    });
+
+    // Update color, painting and random color
+    updatePainting(false);
+    updateColorRandom(getRandomColor(false));
+    updateColorRandomSpinner(getRandomColor(true));
+  }, [fetchQuote, updatePainting, updateColorRandom, updateColorRandomSpinner]);
 
   return (
     <>
@@ -15,18 +47,21 @@ const CardAuthor = ({ contentBtn, stylesBtn, contentCard, event, eventCard, ...p
           <Card.Body>
             <Row>
               <Col>
-                <Card.Title style={{ cursor: "pointer" }} onClick={() => eventCard(title)}>
+                <Card.Title
+                  style={{ cursor: "pointer" }}
+                  onClick={() => searchQuoteAboutAuthor(title)}
+                >
                   {title}
                 </Card.Title>
                 <Card.Subtitle>{subtitle}</Card.Subtitle>
               </Col>
               <Col xs="3">
                 <ButtonNetwork
-                  classGeneral={classBtn}
+                  className={classBtn}
                   content={contentBtn}
                   size={size}
                   variant={variant}
-                  event={event}
+                  event={nextQuoteRandom}
                   params={[false, null, null]}
                 />
               </Col>
@@ -43,7 +78,6 @@ CardAuthor.propTypes = {
   stylesBtn: PropTypes.arrayOf(PropTypes.string).isRequired,
   contentCard: PropTypes.arrayOf(PropTypes.string).isRequired,
   event: PropTypes.func,
-  eventCard: PropTypes.func,
 };
 
 export default CardAuthor;
