+++ src/Errors/HandleError.jsx
@@ -1,14 +1,16 @@
 import PropTypes from "prop-types";
 
-function ErrorFallback({ error, resetErrorBoundary }) {
+const ErrorFallback = ({ error, resetErrorBoundary }) => {
+  const { message } = error;
+
   return (
     <div role="alert">
       <p className="text-danger">Something went wrong:</p>
-      <pre className="text-white">{error.message}</pre>
+      <pre className="text-white">{message}</pre>
       <button onClick={resetErrorBoundary}>Try again</button>
     </div>
   );
-}
+};
 
 ErrorFallback.propTypes = {
   error: PropTypes.instanceOf(Error).isRequired,
