+++ src/pages/Components/CardQuote.jsx
@@ -10,8 +10,9 @@ import { BsTwitter, BsDownload } from "react-icons/bs";
 const CardQuote = () => {
   const { isColorRandomSpinner } = useContext(ContextQuote);
   const { isQuoteAuthor } = useContext(ContextQuoteFetch);
-  const { isLoading, isError, isSuccess, isData } = isQuoteAuthor;
+  const { isLoading, isError, isSuccess, isData, error } = isQuoteAuthor;
 
+  // Handle success and return some components about data
   const handleSuccess = useCallback((data) => {
     const { results } = data;
 
@@ -38,16 +39,19 @@ const CardQuote = () => {
       <ErrorBoundary FallbackComponent={ErrorFallback}>
         {isLoading ? (
           <Container fluid className="d-flex justify-content-center">
+            {/* Render loading if it is loading  */}
             <Loading variant={isColorRandomSpinner} content="Loading .." animation="border" />
           </Container>
         ) : isError ? (
+          // Render error if it is error
           <CardGenerate
-            quote={`““We've had a error.”`}
+            quote={`“We've had an error ${error}”`}
             icons={[<BiError key="error1" />, <BiError key="error2" />]}
             styles={["bg-danger", "fs-4 fst-normal lh-base text-dark", ["mt-4", "md", "dark"]]}
             style={{ maxWidth: "36rem", width: "60vw" }}
           />
         ) : isSuccess ? (
+          // Render success if it is success
           handleSuccess(isData)
         ) : null}
       </ErrorBoundary>
